refactor: refactoring header base on code review
- replace png to svg component.
- remove SafeAreaView
