+++ src/components/index.tsx
@@ -1,1 +1,2 @@
 export * from './Button';
+export * from './Header';
